refactor: 832 - additional refactoring
